Update to latest Forge & added Capabilities for MultiBlocks

+++ src/main/java/elec332/core/config/Configurable.java
@@ -18,9 +18,9 @@ public @interface Configurable {
 
     String comment() default "";
 
-    int minValue() default 0;
+    float minValue() default 0;
 
-    int maxValue() default 0;
+    float maxValue() default 0;
 
     String[] validStrings() default {};
